// src/shared/enums/transaction-type.enum.ts
/**
 * Copyright (c) 2025 Webapp Software Solutions. All rights reserved.
 * This file belongs to Webapp Software Solutions and is proprietary and confidential.
 * Unauthorized copying of this file, via any medium is strictly prohibited.
 * Website: https://webappssoft.com
 */

export enum TransactionType {
  DEPOSIT = 'DEPOSIT',
  WITHDRAW = 'WITHDRAW',
  TRANSFER = 'TRANSFER',
  BONUS = 'BONUS',
  REFUND = 'REFUND',
  PENALTY = 'PENALTY',
  COMMISSION = 'COMMISSION',
  CASHBACK = 'CASHBACK',
}